Update to latest google3 version
* Require C++14 instead of C++11 (abseil-cpp requires it now)
* Require googletest >= 1.10 (for gmock)
  * Location is now specified with -DGOOGLETEST_ROOT
* New classes:
  * S2CellIterator
  * S2CellIteratorJoin
  * S2Coder
* New StableAngle function
* Delete many deprecated functions
* S2Point is now a different class from Vector3_d, not a typedef for it

+++ src/s2/s2builderutil_lax_polygon_layer.cc
@@ -18,12 +18,17 @@
 #include "s2/s2builderutil_lax_polygon_layer.h"
 
 #include <algorithm>
-#include <memory>
 #include <utility>
+#include <vector>
 
-#include "absl/memory/memory.h"
+#include "s2/id_set_lexicon.h"
+#include "s2/s2builder.h"
+#include "s2/s2builder_graph.h"
+#include "s2/s2builder_layer.h"
 #include "s2/s2builderutil_find_polygon_degeneracies.h"
-#include "s2/s2debug.h"
+#include "s2/s2error.h"
+#include "s2/s2lax_polygon_shape.h"
+#include "s2/s2point.h"
 
 using std::vector;
 
@@ -164,7 +169,7 @@ void LaxPolygonLayer::BuildDirected(Graph g, S2Error* error) {
         (degenerate_boundaries == DegenerateBoundaries::DISCARD_HOLES);
     auto degeneracies = s2builderutil::FindPolygonDegeneracies(g, error);
     if (!error->ok()) return;
-    if (degeneracies.size() == g.num_edges()) {
+    if (degeneracies.size() == static_cast<size_t>(g.num_edges())) {
       if (degeneracies.empty()) {
         MaybeAddFullLoop(g, &loops, error);
       } else if (degeneracies[0].is_hole) {
